# frontend/schemas/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SlotMath.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5152/"
    }
  ],
  "paths": {
...
    "/api/runs/measurements/resource-impact": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignEvidenceOfResourceImpactReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/measurements/sampling-design": {
      "post": {
        "tags": [
          "SlotMath.Api"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingDesignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesignEvidenceOfSamplingDesignReport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DesignEstimate": {
        "required": [
          "mean",
          "standardError",
          "interval",
          "exactReference"
        ],
        "type": "object",
        "properties": {
          "mean": {
            "type": "number",
            "format": "double"
          },
          "standardError": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "interval": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NumericInterval"
              }
            ]
          },
          "exactReference": {
            "type": "string"
          }
        }
      },
      "DesignEvidenceOfResourceImpactReport": {
        "required": [
          "authoredInputSha256",
          "runtime",
          "report",
          "retention"
        ],
        "type": "object",
        "properties": {
          "authoredInputSha256": {
            "type": "string"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeProvenance"
          },
          "report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResourceImpactReport"
              }
            ]
          },
          "retention": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DiagnosticRetention"
              }
            ]
          }
        }
      },
      "DesignEvidenceOfSamplingDesignReport": {
        "required": [
          "authoredInputSha256",
          "runtime",
          "report",
          "retention"
        ],
        "type": "object",
        "properties": {
          "authoredInputSha256": {
            "type": "string"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeProvenance"
          },
          "report": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SamplingDesignReport"
              }
            ]
          },
          "retention": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DiagnosticRetention"
              }
            ]
          }
        }
      },
...
      "DiagnosticRetention": {
        "required": [
          "artifactId",
          "retained",
          "note"
        ],
        "type": "object",
        "properties": {
          "artifactId": {
            "type": [
              "null",
              "string"
            ]
          },
          "retained": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "LikelihoodAtom": {
        "required": [
          "id",
          "target",
          "proposal",
          "weight",
          "observed"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "proposal": {
            "type": "string"
          },
          "weight": {
            "type": "string"
          },
          "observed": {
            "type": "integer",
            "format": "int64"
          },
          "stratum": {
            "type": [
              "null",
              "string"
            ]
          },
          "conditionalWeight": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MarkovModelRequest": {
        "required": [
          "transientMatrix",
          "rewards"
        ],
        "type": "object",
        "properties": {
          "transientMatrix": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "rewards": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "initialState": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "stationary": {
            "type": "boolean",
            "default": false
          },
          "costs": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "NumericInterval": {
        "required": [
          "lower",
          "upper",
          "method",
          "assumptions"
        ],
        "type": "object",
        "properties": {
          "lower": {
            "type": "number",
            "format": "double"
          },
          "upper": {
            "type": "number",
            "format": "double"
          },
          "method": {
            "type": "string"
          },
          "assumptions": {
            "type": "string"
          }
        }
      },
...
      "ResourceImpactReport": {
        "required": [
          "stopProbability",
          "retainedReward",
          "referenceReward",
          "omittedReward",
          "retainedDuration",
          "referenceStatus",
          "horizon",
          "assumptions"
        ],
        "type": "object",
        "properties": {
          "stopProbability": {
            "type": "string"
          },
          "retainedReward": {
            "type": "string"
          },
          "referenceReward": {
            "type": [
              "null",
              "string"
            ]
          },
          "omittedReward": {
            "type": [
              "null",
              "string"
            ]
          },
          "retainedDuration": {
            "type": "string"
          },
          "referenceStatus": {
            "type": "string"
          },
          "horizon": {
            "type": "integer",
            "format": "int32"
          },
          "assumptions": {
            "type": "string"
          }
        }
      },
      "ResourceImpactRequest": {
        "required": [
          "reference",
          "horizon"
        ],
        "type": "object",
        "properties": {
          "reference": {
            "$ref": "#/components/schemas/MarkovModelRequest"
          },
          "horizon": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RuntimeProvenance": {
        "required": [
          "measurementContract",
          "numericalMethods",
          "framework",
          "coreBinarySha256",
          "apiBinarySha256"
        ],
        "type": "object",
        "properties": {
          "measurementContract": {
            "type": "string"
          },
          "numericalMethods": {
            "type": "string"
          },
          "framework": {
            "type": "string"
          },
          "coreBinarySha256": {
            "type": [
              "null",
              "string"
            ]
          },
          "apiBinarySha256": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SamplingAtom": {
        "required": [
          "id",
          "value",
          "targetProbability",
          "proposalProbability"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "targetProbability": {
            "type": "string"
          },
          "proposalProbability": {
            "type": "string"
          },
          "stratum": {
            "type": "string",
            "default": "all"
          }
        }
      },
      "SamplingDesignReport": {
        "required": [
          "mode",
          "samples",
          "seed",
          "targetSupportCovered",
          "likelihoods",
          "reward",
          "tailProbability",
          "exactEstimatorVariance",
          "exactWeightSecondMoment",
          "effectiveSampleSize",
          "assumptions"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "samples": {
            "type": "integer",
            "format": "int32"
          },
          "seed": {
            "type": "integer",
            "format": "int64"
          },
          "targetSupportCovered": {
            "type": "boolean"
          },
          "likelihoods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikelihoodAtom"
            }
          },
          "reward": {
            "$ref": "#/components/schemas/DesignEstimate"
          },
          "tailProbability": {
            "$ref": "#/components/schemas/DesignEstimate"
          },
          "exactEstimatorVariance": {
            "type": "string"
          },
          "exactWeightSecondMoment": {
            "type": "string"
          },
          "effectiveSampleSize": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "assumptions": {
            "type": "string"
          }
        }
      },
      "SamplingDesignRequest": {
        "required": [
          "atoms"
        ],
        "type": "object",
        "properties": {
          "atoms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SamplingAtom"
            }
          },
          "mode": {
            "type": "string",
            "default": "importance"
          },
          "samples": {
            "type": "integer",
            "format": "int32",
            "default": 10000
          },
          "seed": {
            "type": "integer",
            "format": "int64",
            "default": 42
          },
          "threshold": {
            "type": "string",
            "default": "1"
          },
          "allocation": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/StratumAllocation"
            }
          },
          "confidence": {
            "type": "number",
            "format": "double",
            "default": 0.95
          },
          "errorFamilySize": {
            "type": "integer",
            "format": "int32",
            "default": 1
          }
        }
      },
...
      "StratumAllocation": {
        "required": [
          "id",
          "samples"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "samples": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}